requirements.txt
streamlit==1.33.0
google-generativeai==0.4.0
python-dotenv==1.1.0
streamlit-authenticator
